Makes Dvara correctly rewrite passive hosts
Currently Dvara does not proxy passive hosts. A passive host is one that
has a priority of 0. This happens when new hosts are added to a replica
set for example.

Certain mongo drivers look at the list of passive hosts to generate its
lists of hosts to connect to, and if Dvara does not proxy these
hosts correctly it can result in strange state being returned to the
driver and causing issues.